requirements.txt
Flask==2.3.2
Flask-Cors==4.0.0
openai==1.14.2
biopython==1.79
transformers==4.18.0
torch==2.2.1
numpy==1.22.3
faiss-cpu==1.7.2
waitress==3.0.0
python-dotenv=1.0.1